+++ tests/config/test_config.py
@@ -10,7 +10,7 @@ from pathlib import Path
 from assertpy import assert_that, soft_assertions
 
 # First party imports
-from printlinter import DEFAULT_IGNORED_REP, Config, OutputLevel
+from printlinter import DEFAULT_IGNORED_REP, Config, OutputLevel, SortedOutput
 
 # Local imports
 from ..conftest import INPUT_FILE_PATH
@@ -22,7 +22,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
 # TODO: Create a config generator and generate X (100?) tests with it.
 @pytest.mark.parametrize(
     "given_file, expect_target_version, expect_ignored_files, expect_disabled_rules, "
-    "expect_color, expect_output_level",
+    "expect_color, expect_output_level, expect_sorted_by",
     [
         # full config
         param(
@@ -32,6 +32,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             ["PPL001"],
             True,
             OutputLevel.L1,
+            SortedOutput.BY_FILES,
             id="yml file",
         ),
         param(
@@ -41,6 +42,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             ["PPL002"],
             False,
             OutputLevel.L2,
+            SortedOutput.BY_ERRORS,
             id="yaml file",
         ),
         param(
@@ -50,6 +52,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             ["PPL003"],
             True,
             OutputLevel.L1,
+            SortedOutput.BY_ERRORS,
             id="json file",
         ),
         param(
@@ -59,6 +62,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             ["PPL004"],
             False,
             OutputLevel.L3,
+            SortedOutput.BY_FILES,
             id="toml file",
         ),
         # partial config
@@ -69,6 +73,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             [],
             True,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial yml file, only target_version",
         ),
         param(
@@ -78,6 +83,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             [],
             True,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial yml file, only ignored_files",
         ),
         param(
@@ -87,6 +93,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             ["PPL001"],
             True,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial yml file, only disabled_rules",
         ),
         param(
@@ -96,6 +103,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             [],
             False,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial yml file, only color",
         ),
         param(
@@ -105,6 +113,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             [],
             True,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial yml file, target_version and ignored_files",
         ),
         param(
@@ -114,6 +123,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             ["PPL001"],
             True,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial yml file, target_version and disabled rules",
         ),
         param(
@@ -123,6 +133,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             ["PPL001"],
             True,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial yml file, ignored_files and disabled_rules",
         ),
         param(
@@ -132,6 +143,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             [],
             False,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial yml file, target_version and color",
         ),
         param(
@@ -141,6 +153,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             [],
             False,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial yml file, target_version, ignored_files and color",
         ),
         param(
@@ -150,6 +163,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             [],
             True,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial yaml file, only target_version",
         ),
         param(
@@ -159,6 +173,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             [],
             True,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial yaml file, only ignored_files",
         ),
         param(
@@ -168,6 +183,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             ["PPL002"],
             True,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial yaml file, only disabled_rules",
         ),
         param(
@@ -177,6 +193,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             [],
             False,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial yaml file, only color",
         ),
         param(
@@ -186,6 +203,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             [],
             True,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial yaml file, target_version and ignored_files",
         ),
         param(
@@ -195,6 +213,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             ["PPL002"],
             True,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial yaml file, target_version and disabled rules",
         ),
         param(
@@ -204,6 +223,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             ["PPL002"],
             True,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial yaml file, ignored_files and disabled_rules",
         ),
         param(
@@ -213,6 +233,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             [],
             False,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial yaml file, target_version and color",
         ),
         param(
@@ -222,6 +243,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             [],
             False,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial yaml file, target_version, ignored_files and color",
         ),
         param(
@@ -231,6 +253,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             [],
             True,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial json file, only target_version",
         ),
         param(
@@ -240,6 +263,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             [],
             True,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial json file, only ignored_files",
         ),
         param(
@@ -249,6 +273,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             ["PPL003"],
             True,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial json file, only disabled_rules",
         ),
         param(
@@ -258,6 +283,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             [],
             False,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial json file, only color",
         ),
         param(
@@ -267,6 +293,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             [],
             True,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial json file, target_version and ignored_files",
         ),
         param(
@@ -276,6 +303,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             ["PPL003"],
             True,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial json file, target_version and disabled rules",
         ),
         param(
@@ -285,6 +313,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             ["PPL003"],
             True,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial json file, ignored_files and disabled_rules",
         ),
         param(
@@ -294,6 +323,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             [],
             False,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial json file, target_version and color",
         ),
         param(
@@ -303,6 +333,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             [],
             False,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial json file, target_version, ignored_files and color",
         ),
         param(
@@ -312,6 +343,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             [],
             True,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial toml file, only target_version",
         ),
         param(
@@ -321,6 +353,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             [],
             True,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial toml file, only ignored_files",
         ),
         param(
@@ -330,6 +363,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             ["PPL004"],
             True,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial toml file, only disabled_rules",
         ),
         param(
@@ -339,6 +373,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             [],
             False,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial toml file, only color",
         ),
         param(
@@ -348,6 +383,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             [],
             True,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial toml file, only target_version and ignored_files",
         ),
         param(
@@ -357,6 +393,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             ["PPL004"],
             True,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial toml file, only target_version and disabled rules",
         ),
         param(
@@ -366,6 +403,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             ["PPL004"],
             True,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial toml file, only ignored_files and disabled_rules",
         ),
         param(
@@ -375,6 +413,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             [],
             False,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial toml file, target_version and color",
         ),
         param(
@@ -384,6 +423,7 @@ TESTING_FILES_PATH = INPUT_FILE_PATH / "config/"
             [],
             False,
             OutputLevel.DEFAULT,
+            SortedOutput.DEFAULT,
             id="partial toml file, target_version, ignored_files and color",
         ),
     ],
@@ -395,6 +435,7 @@ def test_config_from_given_file(
     expect_disabled_rules,
     expect_color,
     expect_output_level,
+    expect_sorted_by,
 ):
     conf = Config(TESTING_FILES_PATH / given_file)
     with soft_assertions():
@@ -403,6 +444,7 @@ def test_config_from_given_file(
         assert_that(conf.disabled_rules).is_equal_to(expect_disabled_rules)
         assert_that(conf.color).is_equal_to(expect_color)
         assert_that(conf.output_level).is_equal_to(expect_output_level)
+        assert_that(conf.sorted_output).is_equal_to(expect_sorted_by)
         assert_that(conf.ignored_rep).is_equal_to(DEFAULT_IGNORED_REP)
 
 
@@ -565,3 +607,8 @@ def test_config_from_file_error_on_target_version(given_file, error):
 def test_config_from_file_error_on_output_level():
     with pytest.raises(ValueError):
         Config(TESTING_FILES_PATH / "errors/output_level_4.yml")
+
+
+def test_config_from_file_error_on_sorted_by():
+    with pytest.raises(ValueError):
+        Config(TESTING_FILES_PATH / "errors/sorted_by_toto.yml")